requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
python-dotenv==0.19.0
Werkzeug==2.0.1